+++ dse_do_utils/scenariodbmanager.py
@@ -57,8 +57,8 @@ class ScenarioDbTable(ABC):
         :param constraints_metadata:
         """
         self.db_table_name = db_table_name
-            # ScenarioDbTable.camel_case_to_snake_case(db_table_name)  # To make sure it is a proper DB table name. Also allows us to use the scenario table name.
-        self.columns_metadata = columns_metadata
+        # ScenarioDbTable.camel_case_to_snake_case(db_table_name)  # To make sure it is a proper DB table name. Also allows us to use the scenario table name.
+        self.columns_metadata = self.resolve_metadata_column_conflicts(columns_metadata)
         self.constraints_metadata = constraints_metadata
         self.dtype = None
         if not db_table_name.islower() and not db_table_name.isupper(): ## I.e. is mixed_case
@@ -69,14 +69,60 @@ class ScenarioDbTable(ABC):
             print(f"Warning: the db_table_name '{db_table_name}' is a reserved word. Do not use as table name.")
         self._sa_column_by_name = None  # Dict[str, sqlalchemy.Column] Will be generated dynamically first time it is needed.
 
+    def resolve_metadata_column_conflicts(self, columns_metadata: List[sqlalchemy.Column]) -> List[sqlalchemy.Column]:
+        columns_dict = {}
+        for column in reversed(columns_metadata):
+            if isinstance(column, sqlalchemy.Column):
+                if column.name in columns_dict:
+                    print(f"Warning: Conflicts in column definition for column {column.name} in table {self.__class__.__name__}. Retained override.")
+                else:
+                    columns_dict[column.name] = column
+            else:
+                print(f"Warning: Column metadata contains non-sqlalchemy in table {self.__class__.__name__}. Retained override.")
+        return list(reversed(columns_dict.values()))
+
     def get_db_table_name(self) -> str:
         return self.db_table_name
 
-    def get_df_column_names(self) -> List[str]:
-        """TODO: what to do if the column names in the DB are different from the column names in the DF?"""
+    def get_df_column_names_2(self, df: pd.DataFrame) -> (List[str], pd.DataFrame):
+        """Get all column names that are defined in the DB schema.
+         If not present in the DataFrame df, adds the missing column with all None values.
+
+         Note 1 (VT 20220829):
+         Note that the `sqlalchemy.insert(db_table.table_metadata).values(row)` does NOT properly handle columns that are missing in the row.
+         It seems to simply truncate the columns if the row length is less than the number of columns.
+         It does NOT match the column names!
+         Thus the need to add columns, so we end up with proper None values in the row for the insert, specifying all columns in the table.
+
+         Note 2 (VT 20220829):
+         Reducing the list of sqlalchemy.Column does NOT work in `sqlalchemy.insert(db_table.table_metadata).values(row)`
+         The db_table.table_metadata is an object, not a List[sqlalchemy.Column]
+
+        :param df:
+        :return:
+        """
         column_names = []
+        # columns_metadata = []
         for c in self.columns_metadata:
             if isinstance(c, sqlalchemy.Column):
+                if c.name in df.columns:
+                    column_names.append(c.name)
+                    # columns_metadata.append(c)
+                else:
+                    column_names.append(c.name)
+                    df[c.name] = None
+
+        return column_names, df
+
+    def get_df_column_names(self, df: pd.DataFrame) -> List[str]:
+        """Get all column names that are both defined in the DB schema and present in the DataFrame df.
+
+        :param df:
+        :return:
+        """
+        column_names = []
+        for c in self.columns_metadata:
+            if isinstance(c, sqlalchemy.Column) and c.name in df.columns:
                 column_names.append(c.name)
         return column_names
 
@@ -138,27 +184,52 @@ class ScenarioDbTable(ABC):
         df.columns = [ScenarioDbTable.camel_case_to_snake_case(x) for x in df.columns]
         return df
 
-    def insert_table_in_db_bulk(self, df: pd.DataFrame, mgr, connection=None):
+    def insert_table_in_db_bulk(self, df: pd.DataFrame, mgr, connection=None
+                                , enable_astype: bool = True
+                                ):
         """Insert a DataFrame in the DB using 'bulk' insert, i.e. with one SQL insert.
         (Instead of row-by-row.)
         Args:
             df (pd.DataFrame)
             mgr (ScenarioDbManager)
             connection: if not None, being run within a transaction
+            enable_astype: if True, apply df.column.astype based on datatypes extracted from columns_metadata (i.e. sqlachemy.Column)
         """
+        if connection is None:
+            connection = mgr.engine
         table_name = self.get_db_table_name()
-        columns = self.get_df_column_names()
+        columns = self.get_df_column_names(df=df)
+        # Note: setting the dtype is NOT equivalent to converting the column using an `astype()`
+        # if enable_dtype:
+        #     dtype = self._get_dtypes()
+        # else:
+        #     dtype = None
+        # print(f"insert_table_in_db_bulk, enable_dtype = {enable_dtype} - dtype = {dtype}")
+
+        # Force the data type of a column in the df to match the expectation in the DB
+        if enable_astype:
+            df = self._set_df_column_types(df)
+
         try:
-            if connection is None:
-                df[columns].to_sql(table_name, schema=mgr.schema, con=mgr.engine, if_exists='append', dtype=None,
+            df[columns].to_sql(table_name, schema=mgr.schema, con=connection, if_exists='append', dtype=None,
                                index=False)
-            else:
-                df[columns].to_sql(table_name, schema=mgr.schema, con=connection, if_exists='append', dtype=None,
-                                   index=False)
         except exc.IntegrityError as e:
             print("++++++++++++Integrity Error+++++++++++++")
             print(f"DataFrame insert/append of table '{table_name}'")
             print(e)
+
+    def _set_df_column_types(self, df: pd.DataFrame) -> pd.DataFrame:
+        """Force the data type of the DataFrame according to the schema.
+        This can avoid errors in DB2, when the column is a mix of types."""
+        for sa_column in self.columns_metadata:
+            df_column_name = sa_column.name
+            df_type = sa_column.type.python_type
+            if type is not None and df_column_name in df.columns:
+                try:
+                    df[df_column_name] = df[df_column_name].astype(df_type)
+                except ValueError as e:
+                    print(f"Failed to convert column {df_column_name} to {df_type}")
+        return df
 
     def _delete_scenario_table_from_db(self, scenario_name, connection):
         """Delete all rows associated with the scenario in the DB table.
@@ -243,13 +314,15 @@ class AutoScenarioDbTable(ScenarioDbTable):
             dtype = ScenarioDbTable.sqlcol(df)
         else:
             dtype = self.dtype
+
+        if connection is None:
+            connection = mgr.engine
+
         try:
             # Note that this can use the 'replace', so the table will be dropped automatically and the defintion auto created
             # So no need to drop the table explicitly (?)
-            if connection is None:
-                df.to_sql(table_name, schema=mgr.schema, con=mgr.engine, if_exists='replace', dtype=dtype, index=False)
-            else:
-                df.to_sql(table_name, schema=mgr.schema, con=connection, if_exists='replace', dtype=dtype, index=False)
+            # TODO: review the 'replace': does it need to be 'append', as in the regular class?
+            df.to_sql(table_name, schema=mgr.schema, con=connection, if_exists='replace', dtype=dtype, index=False)
         except exc.IntegrityError as e:
             print("++++++++++++Integrity Error+++++++++++++")
             print(f"DataFrame insert/append of table '{table_name}'")
@@ -303,7 +376,8 @@ class ScenarioDbManager():
 
     def __init__(self, input_db_tables: Dict[str, ScenarioDbTable], output_db_tables: Dict[str, ScenarioDbTable],
                  credentials=None, schema: str = None, echo: bool = False, multi_scenario: bool = True,
-                 enable_transactions: bool = True, enable_sqlite_fk: bool = True):
+                 enable_transactions: bool = True, enable_sqlite_fk: bool = True, enable_astype: bool = True,
+                 enable_debug_print: bool = False):
         """Create a ScenarioDbManager.
 
         :param input_db_tables: OrderedDict[str, ScenarioDbTable] of name and sub-class of ScenarioDbTable. Need to be in correct order.
@@ -314,12 +388,16 @@ class ScenarioDbManager():
         :param multi_scenario: If true, adds SCENARIO table and PK
         :param enable_transactions: If true, uses transactions
         :param enable_sqlite_fk: If True, enables FK constraint checks in SQLite
+        :param enable_astype: If True, force data-type of DataFrame to match schema before (bulk) insert
+        :param enable_debug_print: If True, print additional debugging statements, like the DB connection string
         """
         # WARNING: do NOT use 'OrderedDict[str, ScenarioDbTable]' as type. OrderedDict is not subscriptable. Will cause a syntax error.
         self.schema = self._check_schema_name(schema)
         self.multi_scenario = multi_scenario  # If true, will add a primary key 'scenario_name' to each table
         self.enable_transactions = enable_transactions
         self.enable_sqlite_fk = enable_sqlite_fk
+        self.enable_astype = enable_astype
+        self.enable_debug_print = enable_debug_print
         self.echo = echo
         self.input_db_tables = self._add_scenario_db_table(input_db_tables)
         self.output_db_tables = output_db_tables
@@ -487,7 +565,8 @@ class ScenarioDbManager():
                 schema=schema
             )
         # SAVE FOR FUTURE LOGGER MESSAGES...
-        #print("Connection String : " + connection_string)
+        if self.enable_debug_print:
+            print("DB2 Connection String : " + connection_string)
         return connection_string
 
     def _create_db2_engine(self, credentials, schema: str, echo: bool = False):
@@ -559,9 +638,8 @@ class ScenarioDbManager():
         See: https://stackoverflow.com/questions/35918605/how-to-delete-a-table-in-sqlalchemy)
         See https://docs.sqlalchemy.org/en/14/core/metadata.html#sqlalchemy.schema.MetaData.drop_all
         """
-        # # Would this work?:
         # print("+++++++++++++++++Reflect+++++++++++++++++")
-        # NOte: the reflect seems to mess things up: when doing the next drop_all we're getting weird errors
+        # Note: the reflect seems to mess things up: when doing the next drop_all we're getting weird errors
         # self.metadata.reflect(bind=connection)  # To reflect any tables in the DB, but not in the current schema
         # print("+++++++++++++++++Drop all tables+++++++++++++++++")
         # self.metadata.drop_all(bind=connection)
@@ -725,14 +803,13 @@ class ScenarioDbManager():
         """
         num_exceptions = 0
         max_num_exceptions = 10
-        columns = db_table.get_df_column_names()
+        columns, df2 = db_table.get_df_column_names_2(df=df)  # Adds missing columns with None values
         #         print(columns)
         # df[columns] ensures that the order of columns in the DF matches that of the SQL table definition. If not, the insert will fail
-        for row in df[columns].itertuples(index=False):
-            #             print(row)
+        for row in df2[columns].itertuples(index=False):
+            # print(row)
             stmt = (
-                sqlalchemy.insert(db_table.table_metadata).
-                    values(row)
+                sqlalchemy.insert(db_table.table_metadata).values(row)
             )
             try:
                 if connection is None:
@@ -1006,8 +1083,8 @@ class ScenarioDbManager():
     # Read multi scenario
     ############################################################################################
     def read_multi_scenario_tables_from_db(self, scenario_names: List[str],
-                                     input_table_names: Optional[List[str]] = None,
-                                     output_table_names: Optional[List[str]] = None) -> (Inputs, Outputs):
+                                           input_table_names: Optional[List[str]] = None,
+                                           output_table_names: Optional[List[str]] = None) -> (Inputs, Outputs):
         """Read selected set input and output tables from multiple scenarios.
         If input_table_names/output_table_names contains a '*', then all input/output tables will be read.
         If empty list or None, then no tables will be read.
@@ -1020,8 +1097,8 @@ class ScenarioDbManager():
         return inputs, outputs
 
     def _read_multi_scenario_tables_from_db(self, connection, scenario_names: List[str],
-                                      input_table_names: List[str] = None,
-                                      output_table_names: List[str] = None) -> (Inputs, Outputs):
+                                            input_table_names: List[str] = None,
+                                            output_table_names: List[str] = None) -> (Inputs, Outputs):
         """Loads data for selected input and output tables from multiple scenarios.
         If either list is names is ['*'], will load all tables as defined in db_tables configuration.
         """
@@ -1124,8 +1201,8 @@ class ScenarioDbManager():
         for scenario_table_name, db_table in self.output_db_tables.items():  # Note this INCLUDES the SCENARIO table!
             if (scenario_table_name != 'Scenario') and scenario_table_name in outputs.keys():  # If in given set of tables to replace
                 df = outputs[scenario_table_name]
-                print(f"Inserting {df.shape[0]} rows and {df.shape[1]} columns in {scenario_table_name}")
-                db_table.insert_table_in_db_bulk(df=df, mgr=self, connection=connection)  # The scenario_name is a column in the df
+                print(f"Inserting {df.shape[0]:,} rows and {df.shape[1]} columns in {scenario_table_name}")
+                db_table.insert_table_in_db_bulk(df=df, mgr=self, connection=connection, enable_astype = self.enable_astype)  # The scenario_name is a column in the df
 
     def replace_scenario_tables_in_db(self, scenario_name, inputs={}, outputs={}):
         """Untested"""
@@ -1152,7 +1229,7 @@ class ScenarioDbManager():
         for scenario_table_name, db_table in self.db_tables.items():  # Note this INCLUDES the SCENARIO table!
             if (scenario_table_name != 'Scenario') and db_table.db_table_name in dfs.keys():  # If in given set of tables to replace
                 df = dfs[scenario_table_name]
-                db_table.insert_table_in_db_bulk(df=df, mgr=self, connection=connection)  # The scenario_name is a column in the df
+                db_table.insert_table_in_db_bulk(df=df, mgr=self, connection=connection, enable_astype = self.enable_astype)  # The scenario_name is a column in the df
 
     ############################################################################################
     # CRUD operations on scenarios in DB: