+++ dse_do_utils/utilities.py
@@ -68,4 +68,20 @@ def list_file_hierarchy(startpath: str) -> None:
             print('{}{}'.format(subindent, f))
 
 
+def convert_size(size_bytes: int):
+    """Returns string describing file size.
+
+    Args:
+        size_bytes (int): size if file in bytes
+
+    From https://stackoverflow.com/questions/5194057/better-way-to-convert-file-sizes-in-python
+    """
+    import math
+    if size_bytes == 0:
+        return "0B"
+    size_name = ("B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB")
+    i = int(math.floor(math.log(size_bytes, 1024)))
+    p = math.pow(1024, i)
+    s = round(size_bytes / p, 2)
+    return "%s %s" % (s, size_name[i])