+++ CHANGELOG.md
@@ -4,9 +4,46 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.5.4.3b]
+## [Unreleased]## [0.5.4.5b3]
+
+## [0.5.4.5b2] - 2022-10-02
+### Changed
+- ScenarioRunner and ScenarioGenerator refactoring
+
+## [0.5.4.5b1] - 2022-09-22
+### Added
+- ScenarioRunner and ScenarioGenerator
+
+## [0.5.4.5b0] - 2022-09-06
+### Changed
+- ScenarioDbManager.resolve_metadata_column_conflicts: resolve conflicts where a ScenarioDbTable subclass redefines a column.
+- ScenarioDbManager._insert_table_in_db_by_row automatically inserts None for columns in defined schema, but missing in DataFrame.
+
+## [0.5.4.4] - 2022-04-28
+### Added
+- ScenarioManager.load_data_from_zip_csv_s: loading input data from a zipped archive of csv files
+- ScenarioManager.write_data_to_zip_csv_s: writing input/output data as a zipped archive of csv files
+### Removed
+- ScenarioManager.write_data_to_csv_s no longer adds files as data assets
+
+## [0.5.4.3] - 2022-04-21
+### Changed
+- Fixed indentation bug in domodeldeployer
+- Update of CONTRIBUTING.md
+- Format thousands for num rows in db-insert progress print message
+### Added
+- DOModelDeployer support for gz/zip package files
+- ScenarioManager.load_data_from_parquet and .write_data_to_parquet
+- ScenarioDbManager.__init__ added parameter `enable_debug_print` to print connection string
+
+## [Unreleased]## [0.5.4.3b0]
 ### Changed
 - Fixed bug in ScenarioDbManager._check_schema_name if schema is None (e.g. when using SQLite)
+- ScenarioDbTable.insert_table_in_db_bulk selects columns present in both the df and the schema. Avoids error when column is defined in DB but not in df.
+### Added
+- Added 'group' as reserved table name
+- Added `local_relative_data_path` in ScenarioManager.__init__() to allow more flexibility in defining local paths
+- ScenarioDbManager.__init__ added parameter `enable_astype: bool = True`: If True, force data-type of DataFrame to match schema before (bulk) insert.
 
 ## [0.5.4.2] - 2022-02-24
 ### Changed